separate connect command =
